handle changing from certificate to configurable

+++ workspaces/mi/mi-extension/src/rpc-managers/mi-visualizer/rpc-manager.ts
@@ -173,7 +173,10 @@ export class MiVisualizerRpcManager implements MIVisualizerAPI {
             await deleteLineFromFile(params.envFilePath, `${params.currentConfigurableName}\n`);
         }
         if (params.currentCertificateFileName && params.currentCertificateFileName !== "") {
-            await deleteFile(params.storedProjectCertificateDirPath + params.currentCertificateFileName);
+            const certificateUsedFiles = params.certificateUsages[params.currentCertificateFileName].resourceUsedFiles;
+            if (!certificateUsedFiles || certificateUsedFiles.length === 1) {
+                await deleteFile(params.storedProjectCertificateDirPath + params.currentCertificateFileName);
+            }
         }
         await appendContent(params.configPropertiesFilePath, `${params.configurableName}:cert\n`);
         await appendContent(params.envFilePath, `${params.configurableName}\n`);
